requirements.txt
streamlit>=1.35.0
pandas~=1.5.3
autogluon==0.8.2
google-cloud-storage>=2.14.0
protobuf~=3.20.2
numpy~=1.26.4
kedro-datasets>=3.0
kedro-viz>=6.7.0
kedro[jupyter]~=0.19.12
seaborn>=0.12.0
scikit-learn~=1.2.2
matplotlib~=3.10.3
Pillow~=9.5.0